Postprocessing

# Example of looking at results at a given time
println("\n==== Simulation done ====")
println("Example of how to postprocess solution and get values at certain times")
analyzeBiofilm(sol,p,[0.26,0.51])

Biofilm.analyzeBiofilm — Function

analyzeBiofilm(sol,p,t)
analyzeBiofilm(sol,p,t,makePlot=true)

Take solution from biofilm solver and outputs variabes and a plot of biofilm variables.